Add inspector method and move tests
